rename uncertainty to nddata
