+++ asdf_astropy/resources/manifests/astropy-1.0.0.yaml
@@ -49,18 +49,3 @@ tags:
     This transform operates on the units of the input, first converting to
     the expected input units, then assigning replacement output units without
     further conversion.
-- tag_uri: tag:astropy.org:astropy/uncertainty/uncertainty-1.0.0
-  schema_uri: http://astropy.org/schemas/astropy/uncertainty/uncertainty-1.0.0
-  title: Represent the nddata.uncertainty.NDUncertainty object
-  description: |-
-    Represents an instance of the nddata.uncertainty.NDUncertainty object
-- tag_uri: tag:astropy.org:astropy/uncertainty/unknownuncertainty-1.0.0
-  schema_uri: http://astropy.org/schemas/astropy/uncertainty/unknownuncertainty-1.0.0
-  title: Represent the nddata.uncertainty.UnknownUncertainty object
-  description: |-
-    Represents an instance of the nddata.uncertainty.UnknownUncertainty object
-- tag_uri: tag:astropy.org:astropy/uncertainty/stddevuncertainty-1.0.0
-  schema_uri: http://astropy.org/schemas/astropy/uncertainty/stddevuncertainty-1.0.0
-  title: Represent the nddata.uncertainty.StdDevUncertainty object
-  description: |-
-    Represents an instance of the nddata.uncertainty.StdDevUncertainty object
